Fixing Build Issues 1.2
