+++ src/main/database/schema/index.ts
@@ -1,5 +1,2 @@
 export * from './user'
-export * from './assignment'
-export * from './academic'
-export * from './log'
-export * from './gradebook'
+export * from './local'
